Improved tracker so it logs out errors (mismatches)

+++ ViewModels/Utils/LogReader.cs
@@ -1,6 +1,7 @@
 ﻿using Lib.DataAccessLayer.Info;
 using Lib.DataAccessLayer.Models;
 using Lib.DataAccessLayer.Repositories;
+using Lib.ViewModels.Services.Dialogs;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
@@ -14,11 +15,19 @@ namespace MainLib.ViewModels.Utils
     public class LogReader
     {
         private string _filesPath;
+        private string _logsPath;
+        private string _path;
         private List<Log<IInfo>> _logs;
+        private List<string> _mismatches;
+
+        public bool NoErrors;
 
         public LogReader(string path)
         {
             this._filesPath = path;
+            this._logsPath = Path.Combine(Environment.CurrentDirectory, "Logs");
+            this.NoErrors = true;
+            this._mismatches = new List<string>();
         }
 
         public void GetLogs(string path)
@@ -29,6 +38,7 @@ namespace MainLib.ViewModels.Utils
 
         public void Sync()
         {
+            // 1. Sync information
             _logs.ForEach((log) =>
             {
                 User user = new UserRepo().GetUserByName(log.Username);
@@ -43,13 +53,21 @@ namespace MainLib.ViewModels.Utils
                             Article article = new Article(local_info);
                             ArticleRepo repo = new ArticleRepo();
 
+                            // Edge case: Article already exists
+                            if (repo.GetArticleWithTitle(article.Title) != null)
+                            {
+                                string mismatch = $"Article '{local_info.Title}' already exists.";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             // 1. Add article to database
                             repo.SaveArticle(article, user);
 
                             // 2. Copy file
                             string fileName = repo.GetFileWithTitle(local_info.Title);
                             File.Copy(
-                                Path.Combine(_filesPath, local_info.FileName + ".pdf"), 
+                                Path.Combine(_filesPath, local_info.FileName + ".pdf"),
                                 Path.Combine(Path.Combine(Environment.CurrentDirectory, "Files"), fileName + ".pdf"));
 
                             Article dbArticle = repo.GetArticleWithTitle(local_info.Title);
@@ -76,12 +94,32 @@ namespace MainLib.ViewModels.Utils
                         else if (log.Info.InfoType == "BookmarkInfo")
                         {
                             BookmarkInfo local_info = log.Info as BookmarkInfo;
-                            new BookmarkRepo().AddBookmark(local_info.Name, local_info.Global, user);
+                            BookmarkRepo repo = new BookmarkRepo();
+
+                            // Edge case: Bookmark already exists
+                            if (repo.GetBookmark(local_info.Name, user) != null)
+                            {
+                                string mismatch = $"Bookmark '{local_info.Name}' already exists.";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
+                            repo.AddBookmark(local_info.Name, local_info.Global, user);
                         }
                         // Create reference
                         else if (log.Info.InfoType == "ReferenceInfo")
                         {
                             ReferenceInfo local_info = log.Info as ReferenceInfo;
+                            ReferenceRepo repo = new ReferenceRepo();
+
+                            // Edge case: Reference already exists
+                            if (repo.GetReference(local_info.Name) != null)
+                            {
+                                string mismatch = $"Reference '{local_info.Name}' already exists.";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             new ReferenceRepo().AddReference(local_info.Name);
                         }
                         break;
@@ -93,16 +131,32 @@ namespace MainLib.ViewModels.Utils
                             ArticleRepo repo = new ArticleRepo();
                             Article existingArticle = repo.GetFullArticleWithTitle(user, log.Changed);
 
+                            // Edge case: Article I am trying to update doesn't exist
+                            if (existingArticle == null)
+                            {
+                                string mismatch = $"Article '{log.Changed}' doesn't exist and can't be updated.";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             Article newArticle = new Article(local_info);
                             newArticle.ID = existingArticle.ID;
                             repo.UpdateArticle(newArticle, user);
                         }
                         // Update bookmark
-                        else if(log.Info.InfoType == "BookmarkInfo")
+                        else if (log.Info.InfoType == "BookmarkInfo")
                         {
                             BookmarkInfo local_info = (BookmarkInfo)log.Info;
                             BookmarkRepo repo = new BookmarkRepo();
                             Bookmark existingBookmark = repo.GetBookmark(log.Changed, user);
+
+                            // Edge case: Bookmark I am trying to update doesn't exist
+                            if (existingBookmark == null)
+                            {
+                                string mismatch = $"Bookmark '{local_info.Name}' doesn't exist and can't be updated.";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
 
                             Bookmark newBookmark = new Bookmark(local_info);
                             newBookmark.ID = existingBookmark.ID;
@@ -114,6 +168,14 @@ namespace MainLib.ViewModels.Utils
                             ReferenceInfo local_info = (ReferenceInfo)log.Info;
                             ReferenceRepo repo = new ReferenceRepo();
                             Reference existingReference = repo.GetReference(log.Changed);
+
+                            // Edge case: Reference I am trying to update doesn't exist
+                            if (existingReference == null)
+                            {
+                                string mismatch = $"Bookmark '{local_info.Name}' doesn't exist and can't be updated.";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
 
                             Reference newReference = new Reference(local_info);
                             newReference.ID = existingReference.ID;
@@ -127,31 +189,95 @@ namespace MainLib.ViewModels.Utils
                         // Couple bookmark
                         if (info.CollectionType == "Bookmark")
                         {
-                            BookmarkRepo repo = new BookmarkRepo();
+                            BookmarkRepo bookmarkRepo = new BookmarkRepo();
+                            ArticleRepo articleRepo = new ArticleRepo();
+
+                            Bookmark bookmark = bookmarkRepo.GetBookmark(info.Name, user);
+                            Article article = articleRepo.GetArticleWithTitle(info.Title);
+
+                            // Edge case: Article or bookmark doesn't exist or Article is already in bookmark
+                            if (bookmark == null)
+                            {
+                                string mismatch = $"Can't couple article - '{info.Title}' with bookmark '{info.Name}' because bookmark doesn't exist";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+                            else if (article == null)
+                            {
+                                string mismatch = $"Can't couple article - '{info.Title}' with bookmark '{info.Name}' because article doesn't exist";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             // Add
                             if (info.ActionType == "Add")
                             {
-                                repo.AddArticleToBookmark(repo.GetBookmark(info.Name, user), new ArticleRepo().GetArticleWithTitle(info.Title));
+                                // Edge case: Article is already in bookmark
+                                if (bookmarkRepo.CheckArticleInBookmark(bookmark, article))
+                                {
+                                    string mismatch = $"Article - '{info.Title}' is already in bookmark '{info.Name}'";
+                                    _mismatches.Add(mismatch);
+                                    return;
+                                }
+                                bookmarkRepo.AddArticleToBookmark(bookmark, article);
                             }
                             // Remove
                             else if (info.ActionType == "Remove")
                             {
-                                repo.RemoveArticleFromBookmark(repo.GetBookmark(info.Name, user), new ArticleRepo().GetArticleWithTitle(info.Title));
+                                // Edge case: Article is not in bookmark
+                                if (!bookmarkRepo.CheckArticleInBookmark(bookmark, article))
+                                {
+                                    string mismatch = $"Article - '{info.Title}' can not be removed from bookmark '{info.Name}' (Its not there)";
+                                    _mismatches.Add(mismatch);
+                                    return;
+                                }
+                                bookmarkRepo.RemoveArticleFromBookmark(bookmark, article);
                             }
                         }
                         // Couple reference
-                        else if(info.CollectionType == "Reference")
+                        else if (info.CollectionType == "Reference")
                         {
-                            ReferenceRepo repo = new ReferenceRepo();
+                            ReferenceRepo referenceRepo = new ReferenceRepo();
+                            ArticleRepo articleRepo = new ArticleRepo();
+
+                            Reference reference = referenceRepo.GetReference(info.Name);
+                            Article article = articleRepo.GetArticleWithTitle(info.Title);
+
+                            // Edge case: Article or bookmark doesn't exist
+                            if (reference == null)
+                            {
+                                string mismatch = $"Can't couple article - '{info.Title}' with reference '{info.Name}' because reference doesn't exist";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+                            else if (article == null)
+                            {
+                                string mismatch = $"Can't couple article - '{info.Title}' with reference '{info.Name}' because article doesn't exist";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             // Add
                             if (info.ActionType == "Add")
                             {
-                                repo.AddArticleToReference(repo.GetReference(info.Name), new ArticleRepo().GetArticleWithTitle(info.Title));
+                                if (referenceRepo.CheckArticleInReference(reference, article))
+                                {
+                                    string mismatch = $"Article - '{info.Title}' is already in reference '{info.Name}'";
+                                    _mismatches.Add(mismatch);
+                                    return;
+                                }
+                                referenceRepo.AddArticleToReference(reference, article);
                             }
                             // Remove
                             else if (info.ActionType == "Remove")
                             {
-                                repo.RemoveArticleFromReference(repo.GetReference(info.Name), new ArticleRepo().GetArticleWithTitle(info.Title));
+                                if (!referenceRepo.CheckArticleInReference(reference, article))
+                                {
+                                    string mismatch = $"Article - '{info.Title}' is already in bookmark '{info.Name}'";
+                                    _mismatches.Add(mismatch);
+                                    return;
+                                }
+                                referenceRepo.RemoveArticleFromReference(reference, article);
                             }
                         }
                         break;
@@ -162,6 +288,15 @@ namespace MainLib.ViewModels.Utils
                         {
                             ArticleRepo repo = new ArticleRepo();
                             Article article = repo.GetArticleWithTitle(local_info1.Name);
+
+                            // Edge case: Article I am trying to delete doesn't exist
+                            if (article == null)
+                            {
+                                string mismatch = $"Can't delete article '{local_info1.Name}' because it doesn't exist";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             string file = repo.GetFileWithTitle(local_info1.Name);
                             repo.DeleteArticle(article);
                             File.Delete(Path.Combine(Environment.CurrentDirectory, "Files", file + ".pdf"));
@@ -171,6 +306,15 @@ namespace MainLib.ViewModels.Utils
                         {
                             BookmarkRepo repo = new BookmarkRepo();
                             Bookmark bookmark = repo.GetBookmark(local_info1.Name, user);
+
+                            // Edge case: Bookmark I am trying to delete doesn't exist
+                            if (bookmark == null)
+                            {
+                                string mismatch = $"Can't delete bookmark '{local_info1.Name}' because it doesn't exist";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             repo.DeleteBookmark(bookmark);
                         }
                         // Delete reference
@@ -178,6 +322,15 @@ namespace MainLib.ViewModels.Utils
                         {
                             ReferenceRepo repo = new ReferenceRepo();
                             Reference reference = repo.GetReference(local_info1.Name);
+
+                            // Edge case: Reference I am trying to delete doesn't exist
+                            if (reference == null)
+                            {
+                                string mismatch = $"Can't delete reference '{local_info1.Name}' because it doesn't exist";
+                                _mismatches.Add(mismatch);
+                                return;
+                            }
+
                             repo.DeleteReference(reference);
                         }
                         break;
@@ -185,6 +338,15 @@ namespace MainLib.ViewModels.Utils
                         PersonalInfo local_info2 = (PersonalInfo)log.Info;
                         ArticleRepo repo1 = new ArticleRepo();
                         Article article3 = repo1.GetArticleWithTitle(local_info2.Title);
+
+                        // Edge case: Article I am trying to add personal doesn't exist
+                        if (article3 == null)
+                        {
+                            string mismatch = $"Can't add personal to article '{local_info2.Title}' because it doesn't exist";
+                            _mismatches.Add(mismatch);
+                            return;
+                        }
+
                         article3.PersonalComment = local_info2.PersonalComment;
                         article3.SIC = local_info2.SIC;
 
@@ -194,6 +356,45 @@ namespace MainLib.ViewModels.Utils
                         break;
                 }
             });
+
+            // 2. Write mismatches
+            WriteMismatches();
+        }
+
+        private void WriteMismatches()
+        {
+            // 1. Set error status
+            this.NoErrors = !(_mismatches.Count > 0);
+
+            if (_mismatches.Count == 0)
+                return;
+
+            // 2. Start writing
+            // File properties
+            string name = "Sync mismatch --- " + DateTime.Today.ToLongDateString() + ".txt";
+            this._path = Path.Combine(_logsPath, name);
+            DateTime currentTime = DateTime.Now;
+
+            //[Obsolete File.Append already creates file if it doesn't exist and this causes another process exception]
+            /*
+            // Create if it doesn't exist
+            if (!File.Exists(_path))
+                File.Create(_path);
+            */
+
+            // Write mismatch into file
+            using (StreamWriter sw = File.AppendText(_path))
+            {
+                // Start writing log starting with current hour as timestamp
+                sw.WriteLine("");
+                sw.WriteLine(currentTime.ToLongTimeString() + ":");
+                sw.WriteLine("");
+
+                _mismatches.ForEach((mismatch) =>
+                {
+                    sw.WriteLine(" - " + mismatch);
+                });
+            }
         }
     }
 }
